# feat(docker): add Dockerfile for FastAPI application and update iac c…