--- a/workshop/src/apps/FastApiApp/Dockerfile
+++ b/workshop/src/apps/FastApiApp/Dockerfile
@@ -1,18 +1,20 @@
 # Use AWS Lambda Python base image with RIC
 FROM public.ecr.aws/lambda/python:3.9
 
+EXPOSE 8000
+
 # Set environment variables for Lambda
+ENV PYTHONPATH=${LAMBDA_TASK_ROOT}
 ENV DEPLOYMENT_TYPE=lambda
-ENV IMAGE_TYPE=custom
-ENV RUNTIME_INTERFACE=ric
+ENV IMAGE_TYPE=public.ecr.aws/lambda/python:3.9
+ENV RUNTIME_INTERFACE=direct
 
 # Copy requirements and install dependencies
-COPY requirements.txt ${LAMBDA_TASK_ROOT}
+COPY workshop/src/apps/FastApiApp/requirements.txt ${LAMBDA_TASK_ROOT}
 RUN pip install --no-cache-dir -r requirements.txt
 
 # Copy application files
-COPY main.py ${LAMBDA_TASK_ROOT}
-COPY lambda_handler.py ${LAMBDA_TASK_ROOT}
+COPY workshop/src/apps/FastApiApp ${LAMBDA_TASK_ROOT}
 
 # Set the CMD to your handler
 CMD ["lambda_handler.lambda_handler"]