--- a/workshop/src/apps/FastApiApp/main.py
+++ b/workshop/src/apps/FastApiApp/main.py
@@ -44,4 +44,4 @@ async def echo_data(data: dict):
 
 # For local development
 if __name__ == "__main__":
-    uvicorn.run(app, host="0.0.0.0", port=8000)
+    uvicorn.run(app, host="0.0.0.0", port=80)