--- a/workshop/src/apps/FastApiApp/main.py
+++ b/workshop/src/apps/FastApiApp/main.py
@@ -1,12 +1,28 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
+from pydantic import BaseModel
 import uvicorn
 import os
 
+# Response models for better Swagger documentation
+class MessageResponse(BaseModel):
+    message: str
+    deployment: str
+
+class HealthResponse(BaseModel):
+    status: str
+    service: str
+
+class EchoResponse(BaseModel):
+    echoed: dict
+    received_at: str
+
 app = FastAPI(
     title="CDK Workshop FastAPI",
-    description="Demo API for CDK Workshop",
-    version="1.0.0"
+    description="Demo API for CDK Workshop - A comprehensive API showcasing different deployment strategies",
+    version="1.0.0",
+    docs_url="/docs",  # Swagger UI endpoint
+    redoc_url="/redoc"  # ReDoc endpoint
 )
 
 # Configure CORS
@@ -18,15 +34,15 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-@app.get("/")
+@app.get("/", response_model=MessageResponse, summary="Root endpoint", description="Returns a welcome message")
 async def root():
     return {"message": "Hello from FastAPI!", "deployment": "workshop"}
 
-@app.get("/health")
+@app.get("/health", response_model=HealthResponse, summary="Health check", description="Returns the health status of the service")
 async def health_check():
     return {"status": "healthy", "service": "fastapi-workshop"}
 
-@app.get("/info")
+@app.get("/info", summary="Service information", description="Returns detailed information about the service and deployment")
 async def get_info():
     return {
         "runtime": "FastAPI",
@@ -38,7 +54,7 @@ async def get_info():
         }
     }
 
-@app.post("/echo")
+@app.post("/echo", response_model=EchoResponse, summary="Echo data", description="Echoes back the provided JSON data")
 async def echo_data(data: dict):
     return {"echoed": data, "received_at": "fastapi-endpoint"}