… workshop/src/apps/FastApiApp/Dockerfile.public