# 🎓 Practical Workshop: Deploy Different API Strategies with Minimal Code Changes

## 🎯 Workshop Objective

This PR demonstrates how to deploy **three different API strategies** using the same CDK infrastructure with **minimal configuration changes**. Perfect for hands-on learning and comparing deployment approaches!

## 🔧 Three Deployment Strategies - One Simple Switch

### **Strategy Selection in `workshop/src/iac/bin/iac.ts`**

```typescript
const dockerfileApi = path.join(
    rootPath,
    // Strategy 1: Custom .NET with package RIC
    // "workshop/src/apps/Minimal.Api/Dockerfile"

    // Strategy 2: Python with AWS Lambda public ECR image  
    // "workshop/src/apps/FastApiApp/Dockerfile"

    // Strategy 3: Python with custom slim image + RIC (ACTIVE)
    "workshop/src/apps/FastApiApp/Dockerfile.ric"
);
```

**📝 Workshop Exercise**: Simply uncomment the desired strategy and deploy!

---

## 📊 Deployment Strategies Comparison

| Strategy | Base Image | Runtime | Cold Start | Best For | One-Line Change |
|----------|------------|---------|------------|----------|----------------|
| **🔵 .NET Minimal API** | `mcr.microsoft.com/dotnet/aspnet:9.0-alpine` | .NET 9.0 | ~10-30s | Enterprise APIs | ✅ Comment change |
| **🟢 FastAPI (AWS Lambda)** | `public.ecr.aws/lambda/python:3.9` | Python 3.9 | ~2-5s | Serverless-first | ✅ Comment change |  
| **🟡 FastAPI (Custom RIC)** | `python:3.9-slim` + RIC | Python 3.9 | ~3-7s | Custom requirements | ✅ Comment change |

---

## 🎯 Strategy Deep Dive

### **Strategy 1: .NET Minimal API** 
```dockerfile
# Multi-stage .NET build with Alpine runtime
FROM mcr.microsoft.com/dotnet/aspnet:9.0-alpine AS base
# Full enterprise-grade API with Swagger, versioning, auth
```
- **Use Case**: Enterprise applications, existing .NET teams
- **Features**: Built-in authentication, API versioning, comprehensive logging
- **Endpoints**: `/health`, `/greetings`, Swagger docs

### **Strategy 2: FastAPI with AWS Lambda Base Image**
```dockerfile  
# AWS-optimized Lambda runtime
FROM public.ecr.aws/lambda/python:3.9
# Direct Lambda integration, no custom RIC needed
```
- **Use Case**: Pure serverless, AWS-native deployments
- **Features**: Optimal cold start, built-in Lambda integration
- **Endpoints**: `/`, `/health`, `/info`, `/echo`

### **Strategy 3: FastAPI with Custom RIC** ⭐ **Currently Active**
```dockerfile
# Custom Python runtime with RIC package
FROM python:3.9-slim
RUN pip install awslambdaric
ENTRYPOINT [ "/usr/local/bin/python", "-m", "awslambdaric" ]
```
- **Use Case**: Custom requirements, hybrid deployments
- **Features**: Full control over base image, custom dependencies
- **Endpoints**: `/`, `/health`, `/info`, `/echo`

---

## 🚀 Workshop Instructions

### **Quick Strategy Switch & Deploy**

1. **Choose Your Strategy** (edit `workshop/src/iac/bin/iac.ts`):
   ```bash
   # Open the file and uncomment your preferred strategy
   code workshop/src/iac/bin/iac.ts
   ```

2. **Deploy in One Command**:
   ```bash
   cd workshop/src/iac
   cdk deploy --all
   ```

3. **Test Your Deployment**:
   ```bash
   # Get API Gateway URL from CDK output
   curl https://your-api-gateway-url/health
   curl https://your-api-gateway-url/info  # Shows deployment strategy!
   ```

### **📱 Compare Deployment Info**

Each strategy exposes its configuration via the `/info` endpoint:

```json
{
  "runtime": "FastAPI",
  "deployment_type": "lambda",
  "container_info": {
    "image_type": "python:3.9-slim",     // Shows current strategy
    "runtime_interface": "ric"           // Shows integration method
  }
}
```

---

## 🎯 Workshop Learning Outcomes

### **💡 Key Lessons**
- **Infrastructure Flexibility**: Same CDK code, different runtimes
- **Strategy Trade-offs**: Cold start vs. control vs. optimization  
- **Container Strategies**: Base image selection impact
- **Serverless Patterns**: Lambda integration approaches

### **🔄 Easy Experimentation**
- Switch between strategies with **one line change**
- Compare performance and behavior side-by-side
- Understand when to choose each approach
- Learn container optimization techniques

### **📈 Real-World Applications**
- **Migration Planning**: Evaluate options before committing
- **Multi-Runtime Support**: Support different team preferences
- **Cost Optimization**: Compare pricing models
- **Performance Tuning**: Test different runtime optimizations

---

## 🎓 Workshop Exercise Ideas

### **🏃‍♂️ Quick Exercises (5 minutes each)**

1. **Deploy Strategy Comparison**:
   ```bash
   # Deploy Strategy 2 (AWS Lambda base)
   # Edit iac.ts, uncomment Dockerfile, deploy
   # Record cold start time and response time
   
   # Switch to Strategy 3 (Custom RIC)  
   # Edit iac.ts, uncomment Dockerfile.ric, deploy
   # Compare performance differences
   ```

2. **Performance Testing**:
   ```bash
   # Test cold starts
   curl -w "@curl-format.txt" https://api-url/health
   
   # Test warm performance  
   for i in {1..10}; do curl https://api-url/info; done
   ```

3. **Cost Analysis**:
   - Deploy same application with different strategies
   - Use AWS Cost Explorer to compare costs after 24 hours
   - Document findings for team discussion

---

## 🔍 Current Configuration

**Active Strategy**: Python with Custom Slim Image + RIC
- **File**: `workshop/src/apps/FastApiApp/Dockerfile.ric`
- **Advantages**: Smaller base image, full control, custom dependencies
- **Trade-offs**: Slightly longer cold start vs. AWS Lambda base image

---

## 🛠️ Technical Implementation

### **Dockerfile Comparison**

**AWS Lambda Base** (`Dockerfile`):
```dockerfile
FROM public.ecr.aws/lambda/python:3.9  # AWS-optimized
CMD ["lambda_handler.lambda_handler"]   # Direct integration
```

**Custom RIC** (`Dockerfile.ric`):
```dockerfile  
FROM python:3.9-slim                              # Smaller base
RUN pip install awslambdaric                      # Add RIC manually
ENTRYPOINT ["/usr/local/bin/python", "-m", "awslambdaric"]  # Custom entry
```

### **Environment Variable Tracking**
Each strategy sets identifying environment variables:
- `DEPLOYMENT_TYPE`: `lambda` | `fargate` | `ecs`
- `IMAGE_TYPE`: Base image identifier
- `RUNTIME_INTERFACE`: `direct` | `ric` | `custom`

---

## 🎉 Workshop Success Metrics

After this workshop, participants will:
- ✅ Understand when to choose each deployment strategy
- ✅ Be able to switch deployment approaches with minimal changes
- ✅ Know how to optimize container images for different use cases  
- ✅ Have hands-on experience with Lambda RIC patterns
- ✅ Understand cost implications of different strategies

**🎓 Perfect for team learning sessions and architecture decision workshops!**